checkCompletes and createNewRound
Also changed Scoreboard to store Double instead of Integer (since score can have 0.5 from draws)

+++ fullstack-backend-proj2/src/main/java/com/codewithcled/fullstack_backend_proj1/model/Round.java
@@ -16,7 +16,7 @@ public class Round {
     private Integer roundNum;
 
     @ElementCollection
-    private Map<Long, Integer> scoreboard;
+    private Map<Long, Double> scoreboard;
 
 
     @OneToMany(mappedBy = "round", cascade = CascadeType.ALL, orphanRemoval = true)
@@ -49,11 +49,11 @@ public class Round {
         return id;
     }
 
-    public Map<Long, Integer> getScoreboard() {
+    public Map<Long, Double> getScoreboard() {
         return scoreboard;
     }
 
-    public void setScoreboard(Map<Long, Integer> scoreboard) {
+    public void setScoreboard(Map<Long, Double> scoreboard) {
         this.scoreboard = scoreboard;
     }
