+++ fullstack-backend-proj2/src/main/java/com/codewithcled/fullstack_backend_proj1/model/Tournament.java
@@ -21,7 +21,7 @@ public class Tournament {
     @CollectionTable(name = "tournament_scoreboard", joinColumns = @JoinColumn(name = "tournament_id"))
     @MapKeyColumn(name = "user_id")
     @Column(name = "score")
-    private Map<Long, Integer> scoreboard;
+    private Map<Long, Double> scoreboard;
     private String date;
     private String status; // the statuses are "open", "ongoing", "completed"
 
@@ -121,11 +121,11 @@ public class Tournament {
     }
 
 
-    public Map<Long, Integer> getScoreboard() {
+    public Map<Long, Double> getScoreboard() {
         return scoreboard;
     }
 
-    public void setScoreboard(Map<Long, Integer> scoreboard) {
+    public void setScoreboard(Map<Long, Double> scoreboard) {
         this.scoreboard = scoreboard;
     }
